fix: fix drag & drop displaying and not allow to drag external entities
