+++ src/components/Sidebar/BetweenFoldersLine.tsx
@@ -2,10 +2,7 @@ import { DragEvent, useCallback, useRef, useState } from 'react';
 
 import classNames from 'classnames';
 
-import {
-  getFolderMoveType,
-  hasDragEventAnyData,
-} from '@/src/utils/app/folders';
+import { getFolderMoveType, hasDragEventAnyData } from '@/src/utils/app/move';
 
 import { FeatureType } from '@/src/types/common';
 import { FolderInterface } from '@/src/types/folder';
